wip-#101 Add missing institution admin privileges && refactor codebase
