+++ composer.json
@@ -4,7 +4,7 @@
     "description": "Authorization Service of Tõlkevärav",
     "keywords": ["framework", "laravel"],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.2",
         "laravel/framework": "^10.0",
         "laravel/sanctum": "^3.2",
@@ -24,6 +24,7 @@
         "psr-4": {
             "App\\": "app/",
             "Database\\Factories\\": "database/factories/",
+            "Database\\Helpers\\": "database/helpers/",
             "Database\\Seeders\\": "database/seeders/"
         }
     },
