+++ database/factories/InstitutionUserFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\Enum\InstitutionUserStatus;
+use App\Enums\InstitutionUserStatus;
 use App\Models\Institution;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
