+++ database/factories/PrivilegeRoleFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\Enum\PrivilegeKey;
+use App\Enums\PrivilegeKey;
 use App\Models\Privilege;
 use App\Models\PrivilegeRole;
 use App\Models\Role;
