+++ tests/Feature/Models/Database/PrivilegeTest.php
@@ -2,7 +2,7 @@
 
 namespace Feature\Models\Database;
 
-use App\Enum\PrivilegeKey;
+use App\Enums\PrivilegeKey;
 use App\Models\Privilege;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Arr;
